refactor: create iframe folder
