+++ src/pages/index.page.tsx
@@ -1,5 +1,5 @@
-import { IframeHideBtn } from '../components/IframeHideBtn';
-import { SendDataForm } from '../components/SendDataForm';
+import { IframeHideBtn } from '../components/iframe/IframeHideBtn';
+import { SendDataForm } from '../components/iframe/SendDataForm';
 import styles from './index.module.css';
 
 const Home = () => {
